crewai_example/pyproject.toml
[project]
name = "crewai-example"
version = "0.1.0"
description = "CrewAI example for customer support"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "create-agent-app",
    "crewai>=0.70.0",
    "langwatch-scenario>=0.7.3",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
]

[tool.uv.sources]
create-agent-app = { path = "../", editable = true } 
